# internal/api/handlers/system/openapi.yaml
...
openapi: 3.0.3
info:
  title: HubPlay API
  description: |
    Self-hosted media server REST API. Powers the React web client, the
    Kotlin Android TV client, and any third-party integration that wants
    to browse a HubPlay catalogue or stream from one.

    ## Authentication

    Two auth flows:

    1. **Password (`/auth/login`)** — Username + password → JWT pair
       (access + refresh). Used by the web client.
    2. **Device code (`/auth/device/*`, RFC 8628)** — TV remotes can't
       type passwords. The device asks the server for a short user code,
       displays it, and polls. The operator approves the code on a
       logged-in browser. Used by the Kotlin TV client and any headless
       client.

    Both produce JWT pairs that are **indistinguishable** thereafter —
    same access-token shape, same refresh path.

    ## Wire format

    All JSON responses are envelope-shaped:

    - **Success**: `{ "data": <payload> }`
    - **Error**: `{ "error": { "code", "message", "hint?", "details?", "request_id" } }`

    The `request_id` correlates server logs; the `code` is a stable
    machine-readable identifier (e.g. `INVALID_CREDENTIALS`,
    `PEER_RATE_LIMITED`); the `message` is human-readable English. Some
    error codes adopt RFC-spec lowercase (e.g. `authorization_pending`
    from RFC 8628). Match on `code` programmatically, surface `message`
    to users via your client's i18n.

    ## Streaming (HLS)

...

    Header format mirrors `Accept-CH`/`Vary`:
...

...

...
servers:
...
paths:
...
  /items/search:
    get:
      tags: [browse]
      summary: Full-text search across item titles + people
      description: |
        Returns the same item shape as `/items` but always FTS-keyed.
        Mirrors the filter surface of `/items` so MediaBrowse can keep
        genre/year/rating active while typing a query — server-side so
        the result set is correct beyond the loaded page.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: library_id, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [movie, series, episode] } }
        - { name: genre, in: query, schema: { type: string } }
        - { name: year_from, in: query, schema: { type: integer, minimum: 1800 } }
        - { name: year_to, in: query, schema: { type: integer, minimum: 1800 } }
        - { name: min_rating, in: query, schema: { type: number, format: float, minimum: 0, maximum: 10 } }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ItemSummary" }
                  total: { type: integer }
...
  /stream/{itemId}/{quality}/index.m3u8:
    parameters:
      - { name: itemId, in: path, required: true, schema: { type: string } }
      - { name: quality, in: path, required: true, schema: { type: string }, description: "Variant identifier (e.g. `1080p`, `720p`)" }
    get:
      tags: [stream]
      summary: HLS variant playlist
      responses:
        "200":
          description: Variant playlist
          content:
            application/vnd.apple.mpegurl:
              schema: { type: string }
  /stream/{itemId}/{quality}/{segment}:
    parameters:
      - { name: itemId, in: path, required: true, schema: { type: string } }
      - { name: quality, in: path, required: true, schema: { type: string } }
      - { name: segment, in: path, required: true, schema: { type: string } }
    get:
      tags: [stream]
      summary: HLS segment
      responses:
        "200":
          description: Transport stream segment
          content:
            video/mp2t:
              schema: { type: string, format: binary }
...
  /stream/{itemId}/subtitles:
    parameters:
      - { name: itemId, in: path, required: true, schema: { type: string } }
    get:
      tags: [stream]
      summary: List embedded subtitle tracks
      responses:
        "200":
          description: Tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SubtitleTrack" }
...
  /channels/{channelId}/stream:
    parameters:
      - { name: channelId, in: path, required: true, schema: { type: string } }
    get:
      tags: [livetv, stream]
      summary: Play a channel (HLS passthrough or 302 to transmux)
      description: |
        Dispatches by upstream format:

        - Upstream is HLS (`*.m3u8`) → server proxies the manifest with
          segment URLs rewritten through `/proxy?url=` so auth + the
          upstream rewriter still apply. `Content-Type: application/vnd.apple.mpegurl`.
        - Upstream is anything else (typically raw MPEG-TS over HTTP) →
          `302 Found` to `/channels/{channelId}/hls/index.m3u8` where a
          per-channel ffmpeg transmux session has been (or will be)
          spawned.
      responses:
        "200":
          description: HLS manifest (passthrough)
          content:
            application/vnd.apple.mpegurl: { schema: { type: string } }
        "302": { description: "Redirected to the transmux HLS endpoint" }
        "404":
          description: Channel inactive or not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/auth/login`, `/auth/refresh`, or `/auth/device/poll`.
        Send as `Authorization: Bearer <access_token>`. Tokens expire in
        15 min by default — refresh proactively or on the first 401.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "Stable machine-readable identifier", example: "INVALID_CREDENTIALS" }
            message: { type: string, description: "Human-readable English; surface via i18n in your client" }
            hint: { type: string, description: "Optional remediation pointer" }
            details: { type: object, additionalProperties: true }
            request_id: { type: string, description: "Correlation id for server logs" }
...
    ItemSummary:
      type: object
      required: [id, type, title]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [movie, series, season, episode] }
        title: { type: string }
        year: { type: integer, nullable: true }
        rating: { type: number, format: float, nullable: true }
        runtime_minutes: { type: integer, nullable: true }
        poster_image_id: { type: string, nullable: true, description: "Use with `/images/file/{id}`" }
        backdrop_image_id: { type: string, nullable: true }
        poster_blurhash: { type: string, nullable: true, description: "BlurHash placeholder for instant render" }
        backdrop_colors:
          type: object
          nullable: true
          properties:
            vibrant: { type: string, example: "rgb(180, 60, 90)" }
            muted: { type: string }
        # Series/season/episode hierarchy
        parent_id: { type: string, format: uuid, nullable: true }
        season_number: { type: integer, nullable: true }
        episode_number: { type: integer, nullable: true }
        # User state (when authenticated)
        user_data:
          type: object
          nullable: true
          properties:
            played: { type: boolean }
            position_seconds: { type: number, format: float }
            is_favorite: { type: boolean }
...
    SubtitleTrack:
      type: object
      properties:
        index: { type: integer }
        codec: { type: string }
        language: { type: string }
        title: { type: string }
        default: { type: boolean }
        forced: { type: boolean }

    # ─── User data ───────────────────────────────────────────────────
...
security:
  - bearerAuth: []